[PDI-12361][PDI-12381] Time To Live (TTL) is not working with Thrift mode.
Execute for each row checkbox is now disabled when using thrift IO in Cassandra input as it doesn't make sense in this case (because a where clause can't be used with thrift mode).
Fixed a UI bug in Cassandra output that prevented the "None" TTL setting to be set once it had first been set to something else.